refactor: remove product/opensource route and add new /opensource route
